requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.11.1
django-environ==0.4.5
gunicorn==20.0.4
Pillow==8.1.0
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pytz==2020.1
sqlparse==0.4.1
